fix(hono,shared): align dist .d.ts with sources and fix server.tsx imports
Two related fixes pulled out of the event-types refactor.

1. integrations/hono/server.tsx was still importing Counter/Form/
   ReactiveProps/PortalExample as default exports after the earlier
   cleanup removed those defaults. Convert to named imports — the
   hono CI e2e job was failing on the missing default exports.

2. packages/adapter-hono shipped a stub `dist/jsx/jsx-runtime/index.d.ts`
   that was missing the JSX namespace and `Async` declaration. The
   hand-written `src/.../index.d.ts` held the real types but `tsgo`
   emitted from the sibling `index.ts` (runtime exports only),
   overwriting the dist copy with a stub. The package.json `types`
   entry pointed at `src/` as a workaround.

   Merge each `index.d.ts` into its `index.ts` using `export declare
   namespace` / `export declare function` — these are ambient so bun
   build emits no runtime, while tsgo now generates a complete dist
   `.d.ts`. Point `types` at the dist copy for consistency with the
   package's other entries.

Audited adapter-go-template, adapter-mojolicious, adapter-tests,
client, jsx, and the rest of `packages/*` for the same overwrite
pattern. Only adapter-hono had paired `.ts` + `.d.ts` files at the
same path; packages/jsx ships type-only `.d.ts` files with no `.ts`
siblings, so tsgo doesn't touch them.
