+++ integrations/hono/server.tsx
@@ -10,13 +10,13 @@ import type { Context } from 'hono'
 import { getCookie, setCookie } from 'hono/cookie'
 import { createDevReloader } from '@barefootjs/hono/dev-worker'
 import { renderer } from './renderer'
-import Counter from '@/components/Counter'
+import { Counter } from '@/components/Counter'
 import Toggle from '@/components/Toggle'
 import TodoApp from '@/components/TodoApp'
 import TodoAppSSR from '@/components/TodoAppSSR'
-import ReactiveProps, { PropsReactivityComparison } from '@/components/ReactiveProps'
-import Form from '@/components/Form'
-import PortalExample from '@/components/PortalExample'
+import { ReactiveProps, PropsReactivityComparison } from '@/components/ReactiveProps'
+import { Form } from '@/components/Form'
+import { PortalExample } from '@/components/PortalExample'
 import ConditionalReturn from '@/components/ConditionalReturn'
 import { AIChatPage } from './components/AIChatPage'
