--- packages/adapter-hono/src/jsx/jsx-dev-runtime/index.d.ts
@@ -1,8 +1,0 @@
-/**
- * BarefootJS Hono JSX Extension - Development Runtime Type Definitions
- *
- * Re-exports JSX namespace from jsx-runtime for development mode.
- */
-
-export { jsxDEV, Fragment } from 'hono/jsx/jsx-dev-runtime'
-export { JSX } from '../jsx-runtime'
